resume_parser_service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pymongo==4.6.0
python-dotenv==1.0.0
spacy==3.7.2
pdfplumber==0.10.3
python-docx==1.0.1
google-cloud-vision==3.4.4
google-cloud-storage==2.13.0
pydantic==2.5.2
pydantic-settings==2.1.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
email-validator==2.1.0.post1